methods/EverCore/pyproject.toml
[project]
name = "memsys"
version = "0.1.0"
description = "Memory System with Dynamic and Static Memory Management"
license = {text = "Apache-2.0"}
requires-python = ">=3.12,<3.13"
dependencies = [
    # Core Dependencies
    "langgraph>=0.2.6",
    "langgraph-api",
    "langgraph-cli",
    "langgraph-sdk>=0.1.57",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langchain>=0.3.26",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.3.28",
    "langchain-anthropic>=0.3.17",
    "langchain-google-genai>=2.1.8",
    "google-genai>=1.21.0",
    # Data Processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    # OpenAI API
    "openai>=1.0.0",
    # Data Validation & Configuration
    "pydantic[email]>=2.7.1",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.0",
    # Web Framework & API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "greenlet>=3.2.0",
    # HTTP Client & File Processing
    "aiohttp>=3.8.0",
    "python-multipart>=0.0.6",
    "aiofiles>=24.1.0",
    # Message Queue
    "aiokafka>=0.11.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Database & Cache
    "redis>=5.0.0",
    "sqlmodel>=0.0.19",
    "asyncpg>=0.29.0",
    "psycopg[binary,pool]>=3.1.0",
    # MongoDB ODM - MongoDB Object Document Mapper
    "beanie>=1.26.0,<3",  # Async MongoDB ODM based on Pydantic (pin major to avoid breaking changes)
    "pymongo>=4.15.4,<5",  # Official MongoDB Python driver (pin major: 5.x has breaking changes)
    # Elasticsearch
    "elasticsearch>=8.17.0,<9",  # ES Python client (pin major: 9.x may change perform_request API)
    "elasticsearch-dsl[async]>=8.17.1,<9",  # ES DSL (pin major, must match elasticsearch major)
    "elastic-transport>=8.17.0,<9",  # ES transport layer (pin major, transport API stability)
    # Milvus
    "pymilvus>=2.5.0,<2.6",  # Pin minor: 2.6.x may change Collection/search API
    # Tokenization
    "jieba==0.42.1",
    # Graph Processing
    "igraph>=0.10.0",
    # Utilities
    "tqdm>=4.65.0",
    "arq>=0.26.0",
    "python-snappy>=0.7.3",
    "tzlocal>=5.3.0",
    "rank-bm25>=0.2.2",
    "duckdb>=1.4.0",
    "import-parent-dir>=0.1.1",
    "bs4>=0.0.2",
    "html2text>=2025.4.15",
    "certifi>=2025.10.5",
    "aiolimiter>=1.2.1",
    "nltk>=3.9.2",
    "tiktoken>=0.12.0",
    "prometheus-client>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']
exclude = '''
(
  /(
      \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | migrations
  | __pycache__
)
'''
include = '\.pyi?$'
skip-string-normalization = true
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests", "unit_test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
# Development dependencies - for local development
dev = [
    "black>=25.1.0",
    "ipdb>=0.13.13",
    "isort>=6.0.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.0.0",
    "nest-asyncio",
    "typer",
    "pyinstrument",
    "py-spy",
    "pre-commit>=4.3.0",
]

cpu = [
]

gpu = [
]

# Full development environment - includes dev tools
dev-full = [
    {include-group = "dev"},
    {include-group = "cpu"},
]

# Production environment - GPU only
prod = [
    {include-group = "gpu"},
]

# Evaluation framework dependencies
evaluation = [
    # Terminal formatting and progress display
    "rich>=13.0.0",
    # HTTP request library (for online API calls)
    "requests>=2.31.0",
]

# Evaluation + all online API clients (full testing environment)
evaluation-full = [
    {include-group = "evaluation"},
    # Online Memory Systems API 客户端
    "mem0ai>=1.0.0",          # Mem0
    "zep-cloud>=2.0.0",       # Zep
]

[project.scripts]
bootstrap = "src.bootstrap:main"
web = "src.run:main"
manage = "src.manage:cli"

[project.entry-points."memsys.addons"]
core = "src.addon"
